+++ packages/shared/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./tenancy.ts";
 export * from "./dinero.ts";
 export * from "./barrio.ts";
+export * from "./liquidacion.ts";